youtube-outline/server/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
langchain-openai==0.1.23
youtube-transcript-api>=0.6.2
pinecone-client==5.0.1
langchain==0.2.15
python-dotenv==1.0.1
openai==1.55.3
httpx<0.28.0
langchain-community==0.2.15
langchain-core==0.2.37
langchain-experimental==0.0.64
langchain-text-splitters==0.2.2
langchainhub==0.1.21
langsmith==0.1.108
langchain-pinecone==0.1.3